nvim: Added list of permanent LSP servers
Currently, whenever I have to configure a fresh working environment, I
have to manually go an install all the LSP servers, often missing some
of them.
This commit configures `Mason` such that it auto-installs[1] the specified
list of LSP (which is and should always be the same as the LSPs we later
configure).

[1]: This requires that `npm` is installed.